chore: rename to TemplateUpdater
